edge collusion detected and some refactoring

+++ src/main/java/model/characters/Hero.java
@@ -11,8 +11,8 @@ public class Hero extends Character
 {
     public Weapon weapon;
     public Armor armor;
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
 
     Hero()
@@ -37,21 +37,21 @@ public class Hero extends Character
         System.out.println(this._name + " is attacking");
     }
 
-    public float getX()
+    public int getX()
     {
         return x;
     }
 
-    public float getY()
+    public int getY()
     {
         return y;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
